+++ pac4j-cas/src/main/java/org/pac4j/cas/redirect/CasRedirectionActionBuilder.java
@@ -7,7 +7,7 @@ import org.pac4j.cas.config.CasConfiguration;
 import org.pac4j.cas.config.CasProtocol;
 import org.pac4j.core.exception.http.RedirectionAction;
 import org.pac4j.core.context.WebContext;
-import org.pac4j.core.exception.http.RedirectionActionHelper;
+import org.pac4j.core.util.HttpActionHelper;
 import org.pac4j.core.redirect.RedirectionActionBuilder;
 import org.pac4j.core.util.CommonHelper;
 import org.slf4j.Logger;
@@ -43,7 +43,7 @@ public class CasRedirectionActionBuilder implements RedirectionActionBuilder {
         final String redirectionUrl = CommonUtils.constructRedirectUrl(computeLoginUrl, getServiceParameter(),
                 computedCallbackUrl, configuration.isRenew(), configuration.isGateway(), configuration.getMethod());
         logger.debug("redirectionUrl: {}", redirectionUrl);
-        return Optional.of(RedirectionActionHelper.buildRedirectUrlAction(context, redirectionUrl));
+        return Optional.of(HttpActionHelper.buildRedirectUrlAction(context, redirectionUrl));
     }
 
     private String getServiceParameter() {