+++ pac4j-http/src/main/java/org/pac4j/http/client/indirect/IndirectBasicAuthClient.java
@@ -4,8 +4,7 @@ import org.pac4j.core.client.IndirectClient;
 import org.pac4j.core.context.HttpConstants;
 import org.pac4j.core.credentials.Credentials;
 import org.pac4j.core.util.Pac4jConstants;
-import org.pac4j.core.exception.http.RedirectionActionHelper;
-import org.pac4j.core.exception.http.UnauthorizedAction;
+import org.pac4j.core.util.HttpActionHelper;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.credentials.authenticator.Authenticator;
 import org.pac4j.core.exception.CredentialsException;
@@ -51,7 +50,7 @@ public class IndirectBasicAuthClient extends IndirectClient {
         assertNotBlank("realmName", this.realmName);
 
         defaultRedirectionActionBuilder(webContext ->
-            Optional.of(RedirectionActionHelper.buildRedirectUrlAction(webContext, computeFinalCallbackUrl(webContext))));
+            Optional.of(HttpActionHelper.buildRedirectUrlAction(webContext, computeFinalCallbackUrl(webContext))));
         defaultCredentialsExtractor(new BasicAuthExtractor());
     }
 
@@ -61,7 +60,7 @@ public class IndirectBasicAuthClient extends IndirectClient {
         assertNotNull("authenticator", getAuthenticator());
 
         // set the www-authenticate in case of error
-        context.setResponseHeader(HttpConstants.AUTHENTICATE_HEADER, "Basic realm=\"" + realmName + "\"");
+        context.setResponseHeader(HttpConstants.AUTHENTICATE_HEADER, HttpConstants.BASIC_HEADER_PREFIX + "realm=\"" + realmName + "\"");
 
         final Optional<Credentials> credentials;
         try {
@@ -70,13 +69,13 @@ public class IndirectBasicAuthClient extends IndirectClient {
             logger.debug("credentials : {}", credentials);
 
             if (!credentials.isPresent()) {
-                throw UnauthorizedAction.INSTANCE;
+                throw HttpActionHelper.buildUnauthenticatedAction(context);
             }
 
             // validate credentials
             getAuthenticator().validate(credentials.get(), context);
         } catch (final CredentialsException e) {
-            throw UnauthorizedAction.INSTANCE;
+            throw HttpActionHelper.buildUnauthenticatedAction(context);
         }
 
         return credentials;