+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/LinkedIn2Client.java
@@ -1,7 +1,7 @@
 package org.pac4j.oauth.client;
 
 import com.github.scribejava.apis.LinkedInApi20;
-import org.pac4j.core.exception.http.RedirectionActionHelper;
+import org.pac4j.core.util.HttpActionHelper;
 import org.pac4j.core.util.CommonHelper;
 import org.pac4j.oauth.exception.OAuthCredentialsException;
 import org.pac4j.oauth.profile.linkedin2.LinkedIn2Configuration;
@@ -50,7 +50,7 @@ public class LinkedIn2Client extends OAuth20Client {
             return false;
         });
         defaultLogoutActionBuilder((ctx, profile, targetUrl) ->
-            Optional.of(RedirectionActionHelper.buildRedirectUrlAction(ctx, "https://www.linkedin.com/uas/logout")));
+            Optional.of(HttpActionHelper.buildRedirectUrlAction(ctx, "https://www.linkedin.com/uas/logout")));
         defaultProfileCreator(new LinkedIn2ProfileCreator(configuration, this));
 
         super.internalInit();