+++ documentation/docs/release-notes.md
@@ -19,6 +19,7 @@ title: Release notes&#58;
 - The `ProfileHelper.restoreOrBuildProfile` method has been removed and the behavior is controlled by the `ProfileDefinition` and its `setRestoreProfileFromTypedId` method (enabled for JWT, disabled for others)
 - Authorizers and matchers can be defined additionaly with "+"
 - Generate CSRF tokens per HTTP request and with an internal expiration date (4 hours)
+- Improved responses for unauthenticated users: 401 with "WWW-Authenticate" header or 403 to be compliant with the HTTP spec
 
 **v4.2.0**: